feat: Add script to shift x-axis to center of graph
The code changes include adding a new script `shift-x-axis-center.py` that demonstrates how to shift the x-axis to the center of the graph. The script plots the graph of the function F(x) = x^2 using matplotlib.

This commit message follows the established convention of using a prefix to indicate the type of change, in this case, "feat" for a new feature. The message is concise and describes the purpose of the code changes.